api_example/requirements.txt
astroid==2.2.4
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.0.0
isort==4.3.11
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
psycopg2==2.7.7
PyJWT==1.7.1
pylint==2.3.1
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
